rename gen-wt to theta, and err to epsilon
gen-wt used to be theta; putting back to how it was.
err used to eps; putting it back but using full name epsilon.

epsilon and theta are not specific to Bogacz 2015, my original source,
but are also used by Friston--e.g. Feldman and Friston 2010--so I'm
using them to convey more quickly to those who know the lit what's
what.